fix initial date bug for SelectDateWidget in RunInfoForm
The initial date value for the RunInfoForm is only set once.
If the application is deployed on 17. April and then
a new RunInfo submited on 18. April then the initial
value for the date is still 17. April.

This bug is fixed by replacing timezone.now() with
the callable object timezone.now