Do not create UserProfiles manually
Manually creating UserProfile instances when saving/creating a User
object causes problems when trying to import data with

python managa.py loaddata

because the unique key constraint gets violated.

For example databaseA has User A with UserProfile AP, then when loading
the data to databaseB, it first imports the User A and then post_save
signal for the User class will be called, which is executing the line
to create a UserProfile object.

Then after all Users has been imported, django tries to import the
UserProfile objects, but exactly here is where the unique constraint
is broken, because our signal already created a UseProfile instance.

It seems like there is no need to create a UseProfile object anyways
and will be automatically created when trying to modify/ save it.

This commit also adds logging messages for when UserProfile instances
are created or saved.