Lower geckodriver to v.20.1
This commit fixes the ResourceWarning issue that recently seems to
appear in every single selenium test. Occasionally this leads to
Segementation fault errors which break the entire test case.

By lowering the geckodriver from v.21.0 to v.20.1 this issue does
seem to not appear anymore.

The Firefox version used is still 'latest', so "Mozilla Firefox 61.0.2"'
for the day of this commit.

Example Error/Warning Messages:
 - Fatal Python error: Segmentation fault
 - tests/selenium/test_shifter.py /home/travis/.travis/job_stages: line 78:  1893 Segmentation fault      (core dumped) PYTHONWARNINGS=all  pytest --cov=.
 - Exception ignored in: <socket.socket fd=25, family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=6, laddr=('127.0.0.1', 49486), raddr=('127.0.0.1', 34059)>
 - ResourceWarning: unclosed <socket.socket fd=20, family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=6, laddr=('127.0.0.1', 44476), raddr=('127.0.0.1', 39971)>